Début transition versLine2D
j’ai essayé de tout changer en Line2D, ça marche pas
j’ai essayé de convertir les Ligne et Line2D ça marche pas mieux…
« rien de particulier » comme dirait l’autre

+++ src/Bille.java
@@ -5,7 +5,7 @@ public abstract class Bille extends Objet {
     public double masse;
     
     
-    public Bille(double ax,double ay,double vx,double vy, Rectangle aframe,String NomImage){
+    public Bille(int ax,int ay,double vx,double vy, Rectangle aframe,String NomImage){
        //* dy=0 au debut, dx=cste (m*g) 
          super(ax,ay,vx,vy,aframe,NomImage);
          masse=1;
